Add gitlab-ci.yml file (#724)
* Add gitlab-ci.yml file

This adds a builder for gitlab.com

Note that this depends on "git describe" to create the tag.
For this to work, you need to create some base tag with something like:
- git tag -n v0.0.1 -m 'Starting version v0.0.1'

After that, well-defined versions can be created by tagging with the
actual version number.

* Add DEP_VERSION and push image as spark-operator

To be conistent, name the output image with spark-operator in the
registry namespace.