fix: remove preemptionPolicy when priority class name is used (#1236)
Starting at kubernetes 1.19, a PriorityClass can now define a
PreemptionPolicy. By default a pod will have a PreemptionPolicy set to
PreemptLowerPriority.

If the PriorityClass defines a different PreemptionPolicy, there will be
a conflict, so the patcher has to remove the PreemptionPolicy on the
pod. By default a PriorityClass will have a PreemptionPolicy set to
PreemptLowerPriority too, so it's safe to always remove it from the pod
when the priorityClassName is specified for the sparkapplication.